+++ src/main/java/com/xiaoleilu/hutool/watch/WatchMonitor.java
@@ -15,7 +15,6 @@ import java.nio.file.WatchEvent.Kind;
 import java.nio.file.WatchKey;
 import java.nio.file.WatchService;
 
-import com.xiaoleilu.hutool.exceptions.WatchException;
 import com.xiaoleilu.hutool.io.IoUtil;
 import com.xiaoleilu.hutool.log.Log;
 import com.xiaoleilu.hutool.log.LogFactory;